update musiccaps and some scripts
