+++ .gitignore
@@ -10,8 +10,3 @@ log/
 *.log
 outputs/
 data/
-
-.gitignore
-examples/vsr_LRS3/scripts/decode_avhubert_vo_vicuna_7b_noself.sh
-examples/asr_librispeech/scripts/decode_hubert_xtralarge_linear_vicuna_7b_copy.sh
-examples/vsr_LRS3/scripts/decode_avhubert_vo_vicuna_7b_copy.sh
