+++ examples/mc_musiccaps/conf/prompt.yaml
