+++ examples/mc_musiccaps/scripts/decode_musicfm_linear_vicuna_7b_10s.sh
