+++ examples/mc_musiccaps/scripts/finetune_musicfm_linear_vicuna_7b_10s.sh
